IA_Training/requirements.txt
pandas
mysql-connector-python
scikit-learn
joblib
python-dotenv
